move simple monitor to integration test
