+++ host/simpleMonitor.go
@@ -18,22 +18,26 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package membership
+package host
 
-import "fmt"
+import (
+	"fmt"
+
+	"github.com/uber/cadence/common/membership"
+)
 
 type simpleMonitor struct {
-	hostInfo  *HostInfo
-	resolvers map[string]ServiceResolver
+	hostInfo  *membership.HostInfo
+	resolvers map[string]membership.ServiceResolver
 }
 
 // NewSimpleMonitor returns a simple monitor interface
-func NewSimpleMonitor(serviceName string, hosts map[string][]string) Monitor {
-	resolvers := make(map[string]ServiceResolver, len(hosts))
+func newSimpleMonitor(serviceName string, hosts map[string][]string) membership.Monitor {
+	resolvers := make(map[string]membership.ServiceResolver, len(hosts))
 	for service, hostList := range hosts {
 		resolvers[service] = newSimpleResolver(service, hostList)
 	}
-	hostInfo := NewHostInfo(hosts[serviceName][0], map[string]string{RoleKey: serviceName})
+	hostInfo := membership.NewHostInfo(hosts[serviceName][0], map[string]string{membership.RoleKey: serviceName})
 	return &simpleMonitor{hostInfo, resolvers}
 }
 
@@ -44,15 +48,15 @@ func (s *simpleMonitor) Start() error {
 func (s *simpleMonitor) Stop() {
 }
 
-func (s *simpleMonitor) WhoAmI() (*HostInfo, error) {
+func (s *simpleMonitor) WhoAmI() (*membership.HostInfo, error) {
 	return s.hostInfo, nil
 }
 
-func (s *simpleMonitor) GetResolver(service string) (ServiceResolver, error) {
+func (s *simpleMonitor) GetResolver(service string) (membership.ServiceResolver, error) {
 	return s.resolvers[service], nil
 }
 
-func (s *simpleMonitor) Lookup(service string, key string) (*HostInfo, error) {
+func (s *simpleMonitor) Lookup(service string, key string) (*membership.HostInfo, error) {
 	resolver, ok := s.resolvers[service]
 	if !ok {
 		return nil, fmt.Errorf("cannot lookup host for service %v", service)
@@ -60,7 +64,7 @@ func (s *simpleMonitor) Lookup(service string, key string) (*HostInfo, error) {
 	return resolver.Lookup(key)
 }
 
-func (s *simpleMonitor) AddListener(service string, name string, notifyChannel chan<- *ChangedEvent) error {
+func (s *simpleMonitor) AddListener(service string, name string, notifyChannel chan<- *membership.ChangedEvent) error {
 	return nil
 }
